requirements.txt
asgiref==3.7.2
Django==5.0.2
django-filter==23.5
django-guardian==2.4.0
djangorestframework==3.14.0
drf-nested-routers==0.93.5
exceptiongroup==1.2.0
iniconfig==2.0.0
Markdown==3.5.2
packaging==23.2
pluggy==1.4.0
Pygments==2.17.2
pytest==8.0.1
pytz==2024.1
PyYAML==6.0.1
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.9.0
